requirements.txt
streamlit
anthropic
beautifulsoup4
requests
python-dotenv
lxml